2ndMVP(~12.18)/backend/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
DateTime==6.0
Deprecated==1.3.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
finance-datareader==0.9.96
fonttools==4.61.0
gunicorn==23.0.0
idna==3.11
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
multipledispatch==1.0.0
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pykrx==1.0.51
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
wrapt==2.0.1
xlrd==2.0.2
zope.interface==8.1.1
